requirements.txt
absl-py==1.0.0
adbutils==0.11.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
apkutils2==1.0.0
asgiref==3.4.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
av==8.1.0
backcall==0.2.0
bert==2.2.0
bert-serving-client==1.10.0
blinker==1.7.0
cachetools==4.2.4
caer==2.0.8
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
cigam==0.0.3
click==8.1.7
colorama==0.4.6
colour==0.1.5
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.6.4
datasets==2.16.1
decorator==5.1.1
deprecation==2.1.0
dill==0.3.7
distlib==0.3.8
Django==4.0
dnspython==2.4.2
docutils==0.18.1
erlastic==2.0.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==1.2.0
faiss-cpu==1.7.4
ffmpeg-python==0.2.0
filelock==3.12.3
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==2.0
fonttools==4.39.3
frozenlist==1.4.1
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
gast==0.4.0
glcontext==2.3.7
glob2==0.7
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.43.0
h5py==3.6.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==6.8.0
importlib-resources==5.12.0
ipython==8.12.0
isosurfaces==0.1.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jdcal==1.4.1
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==24.2.0
Kivy==2.0.0
kivy-deps.angle==0.3.1
kivy-deps.glew==0.3.0
kivy-deps.gstreamer==0.3.3
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
kivymd==0.104.2
kiwisolver==1.4.4
langchain==0.1.5
langchain-community==0.0.18
langchain-core==0.1.19
langsmith==0.0.86
libclang==12.0.0
llvmlite==0.40.1
loguru==0.7.2
manimgl==1.6.1
ManimPango==0.4.3
mapbox-earcut==1.0.1
Markdown==3.3.6
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
moderngl==5.8.2
moderngl-window==2.4.3
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
multipledispatch==0.6.0
multiprocess==0.70.15
mypy==1.0.0
mypy-extensions==1.0.0
networkx==3.1
nh3==0.2.14
nltk==3.8.1
numba==0.57.1
numpy==1.24.4
oauthlib==3.1.1
openai-whisper==20230314
opencv-contrib-python==4.7.0.68
opencv-python==4.8.1.78
openpyxl==3.0.5
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
pinecone-client==3.0.2
pipenv==2023.11.15
pkginfo==1.9.6
platformdirs==4.1.0
prompt-toolkit==3.0.38
protobuf==3.19.1
psutil==5.9.0
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==2.6.0
pydantic_core==2.16.1
pydub==0.25.1
pyelftools==0.29
pygame==2.3.0
pyglet==2.0.5
Pygments==2.14.0
pymongo==4.6.1
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
pynput==1.7.6
pynut==1.11.2
pyodbc==4.0.32
PyOpenGL==3.1.6
pyparsing==3.0.9
PyPDF2==3.0.1
pypiwin32==223
pyrr==0.10.3
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==303
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzbar==0.1.9
pyzmq==25.1.1
readme-renderer==42.0
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.0
requests-toolbelt==1.0.0
retry==0.9.2
rfc3986==2.0.0
rich==13.3.3
rsa==4.8
safetensors==0.4.2
scikit-learn==1.3.0
scipy==1.11.2
scrcpy==0.0.4
screeninfo==0.8.1
seaborn==0.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
skia-pathops==0.7.4
sniffio==1.3.0
SQLAlchemy==2.0.20
sqlparse==0.4.2
stack-data==0.6.2
svgelements==1.9.1
sympy==1.12
tenacity==8.2.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.8.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tiktoken==0.3.1
tokenizers==0.15.1
tomli==2.0.1
torch==2.1.2
torchdata==0.7.1
torchtext==0.16.2
torchvision==0.16.2
torchviz==0.0.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.37.2
twine==4.0.2
typing==3.6.6
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.196
urllib3==2.1.0
validators==0.20.0
virtualenv==20.25.0
virtualenvwrapper-win==1.2.7
wcwidth==0.2.6
Werkzeug==3.0.1
whichcraft==0.6.1
win32-setctime==1.1.0
wrapt==1.13.3
xlrd==1.2.0
xmltodict==0.13.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.6.0
